add static model
