+++ .idea/libraries/Maven__io_micrometer_micrometer_core_1_3_1.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: io.micrometer:micrometer-core:1.3.1">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/micrometer/micrometer-core/1.3.1/micrometer-core-1.3.1.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/io/micrometer/micrometer-core/1.3.1/micrometer-core-1.3.1-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/io/micrometer/micrometer-core/1.3.1/micrometer-core-1.3.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
