+++ .idea/libraries/Maven__org_latencyutils_LatencyUtils_2_0_3.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: org.latencyutils:LatencyUtils:2.0.3">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/latencyutils/LatencyUtils/2.0.3/LatencyUtils-2.0.3.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/latencyutils/LatencyUtils/2.0.3/LatencyUtils-2.0.3-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/latencyutils/LatencyUtils/2.0.3/LatencyUtils-2.0.3-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
