+++ src/main/java/org/ohnlp/n3c/User.java
@@ -1,0 +1,4 @@
+package org.ohnlp.n3c;
+
+public class User {
+}
