+++ src/main/java/org/ohnlp/web/WebServiceController.java
@@ -1,203 +1,32 @@
 package org.ohnlp.web;
 
-import org.apache.uima.UIMAException;
-import org.apache.uima.UIMAFramework;
-import org.apache.uima.analysis_engine.AnalysisEngine;
-import org.apache.uima.analysis_engine.AnalysisEngineDescription;
-import org.apache.uima.analysis_engine.metadata.AnalysisEngineMetaData;
-import org.apache.uima.fit.factory.TypeSystemDescriptionFactory;
-import org.apache.uima.fit.internal.ResourceManagerFactory;
-import org.apache.uima.fit.util.JCasUtil;
-import org.apache.uima.jcas.JCas;
-import org.apache.uima.resource.ResourceInitializationException;
-import org.apache.uima.resource.ResourceManager;
-import org.apache.uima.resource.metadata.ConfigurationParameterSettings;
-import org.apache.uima.resource.metadata.TypeSystemDescription;
-import org.apache.uima.util.InvalidXMLException;
-import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.ohnlp.medtagger.type.ConceptMention;
+import org.ohnlp.n3c.N3CNLPEngine;
 import org.ohnlp.util.BioPortalAPI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
-import java.io.IOException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.Collection;
 import java.util.HashMap;
-
-import static org.apache.uima.fit.factory.AnalysisEngineFactory.createEngineDescription;
-import static org.apache.uima.fit.factory.JCasFactory.createJCas;
 
 @RestController
 public class WebServiceController {
 
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
-    private static TypeSystemDescription tsd;
-    private AnalysisEngine cmAae;
-
+    private static N3CNLPEngine n3CNLPEngine;
     @Value("${bioportal.apikey}")
     private String bioportalAPIKey;
-
-    /**
-     * Trigger UIMA pipeline from the given texts and return only the concept
-     * mentions
-     *
-     * @param docText
-     * @return
-     */
-    HashMap<String, Collection<ConceptMention>> runPipeline(String docText) {
-
-        HashMap<String, Collection<ConceptMention>> annotMap = new HashMap<String, Collection<ConceptMention>>();
-        Collection<ConceptMention> cms = null;
-        try {
-            JCas cmCas = createJCas(tsd);
-            cmCas.setDocumentText(docText);
-            cmAae.process(cmCas);
-            cms = JCasUtil.select(cmCas, ConceptMention.class);
-
-            annotMap.put("cm", cms);
-
-            cmAae.collectionProcessComplete();
-
-        } catch (UIMAException e) {
-            e.printStackTrace();
-        }
-
-        return annotMap;
-    }
-
-    /**
-     * Wrap list of concept mentions into json list to feed into the "entity" js
-     * field.
-     * 
-     * @param cms List of Annotation from UIMA
-     * @return list of JSON array for Brat to display in the html template
-     */
-    static JSONAnnotation generateConceptMentionBratJson(final Collection<ConceptMention> cms) {
-        JSONArray cmList = new JSONArray();
-        JSONArray attribList = new JSONArray();
-
-        int entityIdInt = 1;
-        int attribIdInt = 1;
-
-        if (cms == null) {
-            // Return empty lists
-            return new JSONAnnotation(cmList, attribList);
-        }
-
-        for (ConceptMention cm: cms) {
-
-            // Format: [${ID}, ${TYPE}, [[${START}, ${END}]]]
-            // note that range of the offsets are [${START},${END})
-            // ref:
-            // https://github.com/arne-cl/brat-embedded-visualization-examples/blob/master/minimal-brat-embedded.htm
-
-            JSONArray entityProperties = new JSONArray();
-            String bratEntityId = String.format("T%d", entityIdInt++);
-
-            entityProperties.add(bratEntityId);
-
-            // detectionMethod: "DictionaryLookup"
-            // detectionMethod: "Matched"
-            if (cm.getDetectionMethod().equals("Matched")) {
-                entityProperties.add("Regex");
-            } else if (cm.getDetectionMethod().equals("DictionaryLookup")) {
-                entityProperties.add("Dict");
-            } else {
-                entityProperties.add("Other");
-            }
-            JSONArray spans = new JSONArray();
-            JSONArray tokenBeginEnd = new JSONArray();
-            tokenBeginEnd.add(cm.getBegin());
-            tokenBeginEnd.add(cm.getEnd());
-            spans.add(tokenBeginEnd);
-            entityProperties.add(spans);
-            cmList.add(entityProperties);
-
-            /**
-             * ConceptMention begin: 9 end: 18 detectionMethod: "Matched" normTarget:
-             * "DRY_COUGH" Certainty: "Positive" semGroup: <null> status: "Present"
-             * sentence: Sentence experiencer: "Patient"
-             */
-
-            JSONArray attribProperties = new JSONArray();
-            attribProperties.add(String.format("A%d", attribIdInt++));
-            attribProperties.add("norm");
-            attribProperties.add(bratEntityId);
-            attribProperties.add(cm.getNormTarget());
-            attribList.add(attribProperties);
-
-            JSONArray certaintyProperties = new JSONArray();
-            certaintyProperties.add(String.format("A%d", attribIdInt++));
-            certaintyProperties.add("certainty");
-            certaintyProperties.add(bratEntityId);
-            certaintyProperties.add(cm.getCertainty());
-            attribList.add(certaintyProperties);
-
-            JSONArray statusProperties = new JSONArray();
-            statusProperties.add(String.format("A%d", attribIdInt++));
-            statusProperties.add("status");
-            statusProperties.add(bratEntityId);
-            statusProperties.add(cm.getStatus());
-            attribList.add(statusProperties);
-
-            JSONArray experiencerProperties = new JSONArray();
-            experiencerProperties.add(String.format("A%d", attribIdInt++));
-            experiencerProperties.add("experiencer");
-            experiencerProperties.add(bratEntityId);
-            experiencerProperties.add(cm.getExperiencer());
-            attribList.add(experiencerProperties);
-        }
-
-        return new JSONAnnotation(cmList, attribList);
-    }
-
 
     /**
      * To initialize UIMA resources and type systesms
      */
     public WebServiceController() {
-
-        // TODO: Change hard code path from the file system into resource reference
-        Path ruleDirPath = Paths.get("covid19");
-        System.out.println("IE Rules:\t" + ruleDirPath.toAbsolutePath().toString());
-
-        try {
-            ResourceManager resMgr = ResourceManagerFactory.newResourceManager();
-            AnalysisEngineDescription descN3cTAE = createEngineDescription(
-                    "desc.medtaggerdesc.aggregate_analysis_engine.N3CAggregateTAE");
-            AnalysisEngineMetaData metadata = descN3cTAE.getAnalysisEngineMetaData();
-
-            ConfigurationParameterSettings settings = metadata.getConfigurationParameterSettings();
-            settings.setParameterValue("Resource_dir", ruleDirPath.toString());
-            metadata.setConfigurationParameterSettings(settings);
-
-            // modified MedXNTypes now importing MedTaggerIE types
-            tsd = TypeSystemDescriptionFactory.createTypeSystemDescription("org.ohnlp.medxn.types.MedXNTypes");
-            tsd.resolveImports(resMgr);
-
-            cmAae = UIMAFramework.produceAnalysisEngine(descN3cTAE, resMgr, null);
-
-            assert(cmAae != null);
-        } catch (InvalidXMLException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } catch (ResourceInitializationException e) {
-            e.printStackTrace();
-        }
-
+        n3CNLPEngine = new N3CNLPEngine();
     }
 
     /**
@@ -225,26 +54,25 @@ public class WebServiceController {
 
     /**
      * How to display NLP results .
-     * 
+     *
      * @return the updated view out of index-template
      */
     @PostMapping("/")
     public ModelAndView submit(@ModelAttribute WebInputText webInputText) {
         // To get a list of concept mentions
 
-        HashMap<String, Collection<ConceptMention>> annotMap = runPipeline(webInputText.getDocText());
+        HashMap<String, Collection<ConceptMention>> annotMap = n3CNLPEngine.runPipeline(webInputText.getDocText());
 
         // render template for the fields in index-template
         ModelAndView indexView = new ModelAndView();
         indexView.setViewName("index-template");
         indexView.addObject("input_text", webInputText.getDocText());
         logger.info(String.format("Input for submit \"%s\"", webInputText.getDocText()));
-
         logger.info(String.format("Output map \"%s\"", annotMap));
 
         // get a list of concepts as JSON list to feed into the template
 
-        JSONAnnotation jsAnnot = generateConceptMentionBratJson(annotMap.get("cm"));
+        JSONAnnotation jsAnnot = JSONAnnotation.generateConceptMentionBratJson(annotMap.get("cm"));
 
         indexView.addObject("cmList", jsAnnot.getCmList());
         indexView.addObject("attribList", jsAnnot.getAttribList());
@@ -265,8 +93,8 @@ public class WebServiceController {
         // TODO: check the input 
 
         // get the list of concept metions
-        HashMap<String, Collection<ConceptMention>> annotMap = runPipeline(doc_text);
-        JSONAnnotation jsAnnot = generateConceptMentionBratJson(annotMap.get("cm"));
+        HashMap<String, Collection<ConceptMention>> annotMap = n3CNLPEngine.runPipeline(doc_text);
+        JSONAnnotation jsAnnot = JSONAnnotation.generateConceptMentionBratJson(annotMap.get("cm"));
 
         // build the output data
         JSONObject data = new JSONObject();
